shoreline/requirements.txt
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
flask-restplus==0.13.0
flask-restx==0.5.1
requests~=2.27.1
six~=1.11.0
PyJWT==1.7.1
pymongo==4.1.1
itsdangerous==2.0.1